// src/lib/api/generated/main/models/postProjectAuthorizationMemberAuthorization.ts
/**
 * Generated by orval v7.5.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type { PostProjectAuthorizationMemberAuthorizationMemberDivision } from './postProjectAuthorizationMemberAuthorizationMemberDivision';

export interface PostProjectAuthorizationMemberAuthorization {
  memberId?: number;
  projectAuthorization?: string;
  memberDivision?: PostProjectAuthorizationMemberAuthorizationMemberDivision;
}